refactor: extract player hotkeys into gesture/keyboard hooks
